+++ bin/src/calc.dart
@@ -1,44 +1,60 @@
 library dart_coveralls.calc;
 
+import 'dart:async';
 import 'dart:io';
 
 import 'package:dart_coveralls/dart_coveralls.dart';
+import 'package:path/path.dart' as p;
 
 import 'command_line.dart';
 
 class CalcPart extends CommandLinePart {
-  final ArgParser parser;
+  CalcPart() : super(_initializeParser());
 
-  CalcPart() : parser = _initializeParser();
+  Future execute(ArgResults res) async {
+    String packageRoot = res["package-root"];
+    if (p.isRelative(packageRoot)) {
+      packageRoot = p.absolute(packageRoot);
+    }
 
-  static ArgParser _initializeParser() {
-    return new ArgParser(allowTrailingOptions: true)
-      ..addFlag("help", help: "Prints this help", negatable: false)
-      ..addOption("workers",
-          help: "Number of workers for parsing", defaultsTo: "1")
-      ..addOption("output", help: "Output file path")
-      ..addOption("package-root",
-          help: "Root of the analyzed package", defaultsTo: ".");
-  }
+    if (!FileSystemEntity.isDirectorySync(packageRoot)) {
+      print("Package root directory does not exist");
+      return;
+    }
 
-  execute(ArgResults res) {
-    if (res.rest.length != 1) return print("Please specify a test file to run");
+    if (res.rest.length != 1) {
+      print("Please specify a test file to run");
+      return;
+    }
 
-    var pRoot = new Directory(res["package-root"]);
-    var file = new File(res.rest.single);
+    var file = res.rest.single;
+    if (p.isRelative(file)) {
+      file = p.absolute(file);
+    }
+
+    if (!FileSystemEntity.isFileSync(file)) {
+      print("Dart file does not exist");
+      return;
+    }
+
     var workers = int.parse(res["workers"]);
+    var collector = new LcovCollector(packageRoot, file);
 
-    if (!pRoot.existsSync()) return print("Root directory does not exist");
-    if (!file.existsSync()) return print("Dart file does not exist");
+    var r = await collector.getLcovInformation(workers: workers);
 
-    var collector = new LcovCollector(pRoot, file);
-
-    return collector.getLcovInformation(workers: workers).then((r) {
-      print(r.processResult.stdout);
-      if (res["output"] != null) {
-        return new File(res["output"]).writeAsStringSync(r.result);
-      }
-      return print(r.result);
-    });
+    r.printSummary();
+    if (res["output"] != null) {
+      await new File(res["output"]).writeAsString(r.result);
+    } else {
+      print(r.result);
+    }
   }
 }
+
+ArgParser _initializeParser() => new ArgParser(allowTrailingOptions: true)
+  ..addFlag("help", help: "Prints this help", negatable: false)
+  ..addOption("workers", help: "Number of workers for parsing", defaultsTo: "1")
+  ..addOption("output", help: "Output file path")
+  ..addOption("package-root",
+      help: 'Where to find packages, that is, "package:..." imports.',
+      defaultsTo: "packages");