+++ lib/src/cli_client.dart
@@ -1,34 +1,46 @@
 library dart_coveralls.cli_client;
 
 import 'dart:async' show Future, Completer;
+import 'dart:convert';
 import 'dart:io';
 
-import 'package:mockable_filesystem/filesystem.dart';
+import 'package:path/path.dart' as p;
 import 'package:stack_trace/stack_trace.dart';
 
 import 'collect_lcov.dart';
-import 'coveralls_entities.dart';
 import 'coveralls_endpoint.dart';
+import 'coveralls_entities.dart';
+import 'log.dart';
 import 'process_system.dart';
 
 class CommandLineClient {
-  final FileSystem fileSystem;
-  final Directory packageRoot;
+  final String projectDirectory;
+  final String packageRoot;
   final String token;
   final String serviceName;
 
-  CommandLineClient(Directory packageRoot, {String token,
-      FileSystem fileSystem: const FileSystem(),
-      Map<String, String> environment})
-      : fileSystem = fileSystem,
-        packageRoot = packageRoot,
-        serviceName = getServiceName(environment),
-        token = getToken(token, environment);
+  CommandLineClient._(
+      this.projectDirectory, this.packageRoot, this.token, this.serviceName);
 
-  Future<CoverageResult<String>> getLcovResult(File testFile,
+  factory CommandLineClient({String projectDirectory, String packageRoot,
+      String token, Map<String, String> environment}) {
+    if (projectDirectory == null) {
+      projectDirectory = p.current;
+    }
+
+    packageRoot = _calcPackageRoot(projectDirectory, packageRoot);
+
+    var serviceName = getServiceName(environment);
+    token = getToken(token, environment);
+
+    return new CommandLineClient._(
+        projectDirectory, packageRoot, token, serviceName);
+  }
+
+  Future<CoverageResult<String>> getLcovResult(String testFile,
       {int workers, ProcessSystem processSystem: const ProcessSystem()}) {
-    var collector = new LcovCollector(packageRoot, testFile,
-        fileSystem: fileSystem, processSystem: processSystem);
+    var collector =
+        new LcovCollector(packageRoot, testFile, processSystem: processSystem);
     return collector.getLcovInformation(workers: workers);
   }
 
@@ -51,23 +63,30 @@ class CommandLineClient {
     return serviceName;
   }
 
-  Future reportToCoveralls(File testFile, {int workers,
+  Future reportToCoveralls(String testFile, {int workers,
       ProcessSystem processSystem: const ProcessSystem(),
       String coverallsAddress, bool dryRun: false,
       bool throwOnConnectivityError: false, int retry: 0,
-      bool excludeTestFiles: false}) async {
+      bool excludeTestFiles: false, bool printJson}) async {
     var rawLcov = await getLcovResult(testFile,
         workers: workers, processSystem: processSystem);
 
-    print(rawLcov.processResult.stdout);
+    rawLcov.printSummary();
     var lcov = LcovDocument.parse(rawLcov.result.toString());
-    var report = CoverallsReport.parse(token, lcov, packageRoot, serviceName,
+    var report = CoverallsReport.parse(
+        token, lcov, projectDirectory, serviceName,
         excludeTestFiles: excludeTestFiles);
     var endpoint = new CoverallsEndpoint(coverallsAddress);
+
+    if (printJson) {
+      print(const JsonEncoder.withIndent('  ').convert(report));
+    }
+
     if (dryRun) return;
 
     try {
-      await _sendLoop(endpoint, report.covString(), retry: retry);
+      var encoded = JSON.encode(report);
+      await _sendLoop(endpoint, encoded, retry: retry);
     } catch (e, stack) {
       if (throwOnConnectivityError) rethrow;
       stderr.writeln('Error sending results');
@@ -77,17 +96,34 @@ class CommandLineClient {
   }
 }
 
+String _calcPackageRoot(String packageDir, String packageRoot) {
+  assert(p.isAbsolute(packageDir));
+
+  if (packageRoot == null) {
+    packageRoot = 'packages';
+  }
+
+  if (p.isRelative(packageRoot)) {
+    packageRoot = p.join(packageDir, packageRoot);
+  }
+
+  return p.normalize(packageRoot);
+}
+
 Future _sendLoop(CoverallsEndpoint endpoint, String covString,
     {int retry: 0}) async {
+  var currentRetryCount = 0;
   while (true) {
     try {
       await endpoint.sendToCoveralls(covString);
       return;
-    } catch (_) {
-      if (retry <= 0) {
+    } catch (e) {
+      if (currentRetryCount >= retry) {
         rethrow;
       }
-      retry--;
+      currentRetryCount++;
+      log.warning('Error sending', e);
+      log.info("Retrying $currentRetryCount of $retry.");
     }
   }
 }