+++ lib/src/collect_lcov.dart
@@ -4,46 +4,9 @@ import "dart:async" show Future;
 import "dart:io";
 
 import "package:coverage/coverage.dart";
-import "package:mockable_filesystem/filesystem.dart" show FileSystem;
-import "package:path/path.dart";
+import "package:path/path.dart" as p;
 
 import "process_system.dart";
-
-/// Checks the given directory for the most recently changed dart coverage file
-File _getYoungestDartCoverageFile(Directory dir) {
-  var files = dir.listSync();
-  var coverageFiles = files
-      .where(_isDartCoverageEntity)
-      .map((f) => f as File)
-      .toList() as List<File>;
-  return _youngestElement(coverageFiles);
-}
-
-/// Returns the most recently changed file among the given files
-File _youngestElement(Iterable<File> files) {
-  if (files.isEmpty) throw new Exception("Empty iterable");
-
-  var max = files.first;
-  var maxStats = max.statSync();
-  var iterable = files.skip(1);
-
-  for (var file in iterable) {
-    var stats = file.statSync();
-    if (0 > stats.changed.compareTo(maxStats.changed)) {
-      maxStats = stats;
-      max = file;
-    }
-  }
-
-  return max;
-}
-
-/// Checks if the given entity is a file and checks its name against a pattern
-bool _isDartCoverageEntity(FileSystemEntity file) {
-  if (file is! File) return false;
-  var name = basename(file.path);
-  return name.startsWith("dart-cov") && name.endsWith(".json");
-}
 
 class LcovDocument {
   final List<LcovPart> parts;
@@ -87,48 +50,62 @@ class LcovPart {
 
 class LcovCollector {
   final String sdkRoot;
-  final File testFile;
-  final Directory packageRoot;
-  final FileSystem fileSystem;
+  final String testFile;
+  final String packageRoot;
   final ProcessSystem processSystem;
 
   LcovCollector(this.packageRoot, this.testFile,
-      {this.fileSystem: const FileSystem(),
-      this.processSystem: const ProcessSystem(), this.sdkRoot});
+      {this.processSystem: const ProcessSystem(), this.sdkRoot}) {
+    if (!p.isAbsolute(testFile)) {
+      throw new ArgumentError.value(
+          testFile, 'testFile', 'Must be an absolute path.');
+    }
+  }
 
+  // TODO: perhaps provide an option to NOT delete the temp file and instead
+  //       print out the path for other tooling
   /// Returns an LCOV string of the tested [File].
   ///
   /// Calculates and returns LCOV information of the tested [File].
   /// This uses [workers] to parse the collected information.
-  Future<CoverageResult<String>> getLcovInformation({int workers: 1}) {
-    var reportFile = _getCoverageJson();
-    return parseCoverage([reportFile.result], workers).then((hitmap) {
-      var resolver =
-          new Resolver(packageRoot: packageRoot.path, sdkRoot: sdkRoot);
+  Future<CoverageResult<String>> getLcovInformation({int workers: 1}) async {
+    Directory tempDir =
+        await Directory.systemTemp.createTemp('dart_coveralls.');
+    try {
+      var reportFile = await _getCoverageJson(tempDir);
+
+      var hitmap = await parseCoverage(reportFile.result, workers);
+      var resolver = new Resolver(packageRoot: packageRoot, sdkRoot: sdkRoot);
       var formatter = new LcovFormatter(resolver);
-      reportFile.result.deleteSync();
-      return formatter
-          .format(hitmap)
-          .then((res) => new CoverageResult(res, reportFile.processResult));
-    });
+
+      var res = await formatter.format(hitmap);
+      return new CoverageResult<String>(res, reportFile.processResult);
+    } finally {
+      await tempDir.delete(recursive: true);
+    }
   }
 
   /// Generates and returns a coverage json file
-  CoverageResult<File> _getCoverageJson() {
-    var current = fileSystem.getDirectory(fileSystem.currentDirectory);
+  Future<CoverageResult<List<File>>> _getCoverageJson(
+      Directory coverageDir) async {
     var args = [
-      "--enable-vm-service:9999",
-      "--coverage_dir=${current.path}",
-      testFile.absolute.path
+      "--coverage_dir=${coverageDir.path}",
+      "--package-root=${packageRoot}",
+      testFile
     ];
-    var result = processSystem.runProcessSync("dart", args);
+    var result = await processSystem.runProcess(Platform.executable, args);
     if (result.exitCode < 0) {
-      throw new ProcessException('dart', args,
+      stderr.writeln('stdout:');
+      stderr.writeln(result.stdout);
+      stderr.writeln('stderr:');
+      stderr.writeln(result.stderr);
+      throw new ProcessException(Platform.executable, args,
           'There was a critical error. Exit code: ${result.exitCode}',
           result.exitCode);
     }
-    var reportFile = _getYoungestDartCoverageFile(current);
-    return new CoverageResult<File>(reportFile, result);
+    var reportFiles =
+        await coverageDir.list(recursive: false, followLinks: false).toList();
+    return new CoverageResult<List<File>>(reportFiles, result);
   }
 }
 
@@ -137,4 +114,16 @@ class CoverageResult<E> {
   final ProcessResult processResult;
 
   CoverageResult(this.result, this.processResult);
+
+  /// Prints `processResult.stdout`
+  ///
+  /// If `processResult.exitCode` is not zero, also prints the exit code and
+  /// `processResult.stderr`.
+  void printSummary() {
+    print(processResult.stdout);
+    if (processResult.exitCode != 0) {
+      print("Process exited with code ${processResult.exitCode}.");
+      print(processResult.stderr);
+    }
+  }
 }